pyproject.toml
[project]
name = "quiz-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "openai-agents>=0.6.1",
    "pypdf>=6.4.0",
]
